// src/enums/cultureCard.ts
export enum CultureCard {
	GloryToTheDespot = 'gloryToTheDespot',
	Uprising = 'uprising',
	BarbarianCamp = 'barbarianCamp',
	Migrants = 'migrants',
	BreadAndCircuses = 'breadAndCircuses',
	GiftFromAfar = 'giftFromAfar',
	Drought = 'drought',
	DutyAndHonor = 'dutyAndHonor',
	ExchangeOfIdeas = 'exchangeOfIdeas',
	Defectors = 'defectors',
	HumanitarianAid = 'humanitarianAid',
	Nationalism = 'nationalism',
	FinancialCrisis = 'financialCrisis',
	Immigrants = 'immigrants',
	OffCourse = 'offCourse',
	Catastrophe = 'catastrophe',
	Deforestation = 'deforestation',
	EpidemicOfRenegades = 'epidemicOfRenegades',
	Rebellion = 'rebellion',
	GodSaveTheQueen = 'godSaveTheQueen',
	JointResearch = 'jointResearch',
	ElementalRampage = 'elementalRampage',
	PrimeTime = 'primeTime',
	BankingScandal = 'bankingScandal',
	RoyalGift = 'royalGift',
	RedeploymentIssues = 'redeploymentIssues',
	Patriotism = 'patriotism',
	PresidentDay = 'presidentDay',
	BlindedWanderings = 'blindedWanderings',
	Counterfeiters = 'counterfeiters',
	Sabotage = 'sabotage',
	AccessToKnowledge = 'accessToKnowledge',
	MassDesertion = 'massDesertion',
	Colonists = 'colonists',
	NomadHorde = 'nomadHorde',
	Joust = 'joust',
	GenerousGift = 'generousGift',
	Flood = 'flood',
}